build: rename build to build-js
To stay in sync with our other repositories.
