+++ build-js/WebpackSPDXPlugin.js
