+++ package.json
@@ -6,7 +6,7 @@
 	"private": true,
 	"scripts": {
 		"build": "NODE_ENV=production webpack --progress --config webpack.js",
-		"postbuild": "build/npm-post-build.sh",
+		"postbuild": "build-js/npm-post-build.sh",
 		"dev": "NODE_ENV=development webpack --progress --config webpack.js",
 		"watch": "NODE_ENV=development webpack --progress --watch --config webpack.js",
 		"lint": "eslint --ext .js,.vue src",
